fix forgotten .show() call and rename ProgressBar
